fastapi_medical_app/fastapi_medical_app/requirements.txt
fastapi
pymongo
python-dotenv
bcrypt
pyjwt
uvicorn